reduced the knowledge of settings service of outside world, by using the minimal interface for requests, rather than playwright context, which it should not be aware of
